src/main/java/pro/sky/skyprospringdemo/service/CalculatorServiceImpl.java
package pro.sky.skyprospringdemo.service;

import org.springframework.stereotype.Service;

@Service
public class CalculatorServiceImpl implements CalculatorService{

    Integer  a = 5;
    Integer  b = 7;

    public void setA(Integer a) {
        this.a = a;
    }

    public void setB(Integer b) {
        this.b = b;
    }

    public Integer getA() {
        return a;
    }

    public Integer getB() {
        return b;
    }

    @Override
    public int plus(Integer a, Integer b) {
        return a + b;
    }

    @Override
    public int minus(Integer a, Integer b) {
        return a - b;
    }

    @Override
    public int multiple(Integer a, Integer b) {
        return a * b;
    }

    @Override
    public double divide(Integer a, Integer b) {
        if(b == 0) throw new IllegalArgumentException("Делить на ноль нельзя!");
               return (double) a / b ;
    }
}
